input_(): Add Windows Unicode support
Simply setting the input code page doesn't help so we
have to use ReadConsoleW. In the fallback case
we decode using the current output codepage. Not sure
if that is correct, but "echo" for example uses the
current output codepage, and this makes piping echo
output to python work.

This also improves the case where setting the utf-8
code page in print_() fails and encodes using the
original code page instead.