uri2fsn: fix handling of relative paths with newer Python
Python 3.12 changed urlunparse() results for relative file uris as a side
effect of some other changes. It's not clear what and why from the
git history, so try to reproduce the new behaviour for all Python
versions (if path doesn't start with / then strip them from the result).

And since urlunparse() is no longer guaranteed to return uris starting
// move the netloc hack up to simplify things.

tbh I'm a bit out of the loop re URIs, so I hope on the test coverage
to prevent any oversights.

Fixes #16